requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
cffi==1.12.3
Click==7.0
cryptography==2.7
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
python-dotenv==0.10.3
six==1.12.0
typed-ast==1.4.0
Werkzeug==0.15.5
wrapt==1.11.2
